modify data structure "configs"
This commit is for resetting rx4sx.scm so it will be reverted soon.

<configs>
configs is a pair of future configurations and past configurations.
car: list of uncalculated configutations
cdr: list of calculated configurations

This data structure is no use because acc? (in conf-accepts?) gets rid
of past configurations of given configs in line 83:

83  ((conf-accepts? M (car-configs configs)) #t)

car-configs returns no information about what is calculated before, so
conf-accepts? can't make use of it. Poor conf-accepts? goes round in
circles on the same string and states.